Move VariableSelection into separate reuseable component.
