+++ src/plugins/graph/components/legend/variable-selection.scss
